+++ Sources/CATempo.swift
@@ -46,8 +46,9 @@ public class CATempo : Tempo {
      */
     public override init() {
         super.init()
-        displayLink = CADisplayLink.init(target: self, selector: #selector(update))
-        displayLink?.add(to: RunLoop.main, forMode: RunLoop.Mode(rawValue: RunLoop.Mode.common.rawValue))
+        
+        displayLink = CADisplayLink(weakTarget: self, selector: #selector(update))
+        displayLink?.add(to: RunLoop.main, forMode: .common)
     }
     
     deinit {