+++ Sources/MotionSequence.swift
@@ -261,7 +261,7 @@ public class MotionSequence: Moveable, MoveableCollection, TempoDriven, MotionUp
     private(set) public var motionState: MotionState
     
     /// Provides a delegate for updates to a Moveable object's status, used by Moveable collections.
-    public var updateDelegate: MotionUpdateDelegate?
+    public weak var updateDelegate: MotionUpdateDelegate?
     
     
     // MARK: MoveableCollection protocol properties
@@ -485,7 +485,7 @@ public class MotionSequence: Moveable, MoveableCollection, TempoDriven, MotionUp
      *      - steps: An array of `Moveable` objects the MotionSequence should control. The positions of the objects in the Array will determine the order in which the child motions should move.
      *      - options: An optional set of `MotionsOptions`.
      */
-    public init(steps: [Moveable] = [], options: MotionOptions? = .none) {
+    public init(steps: [Moveable] = [], options: MotionOptions? = MotionOptions.none) {
         
         // unpack options values
         if let unwrappedOptions = options {