+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.0.1
+- Fixed some retain cycles that were holding on to target objects
+- Updated examples project
+
 #### 2.0.0
 - Support for Swift 5.0
 - Updated syntax in MotionOptions for newer Swift naming conventions